+++ worlds/chibi_robo/rules.py
@@ -29,6 +29,8 @@ def set_rules(self) -> None:
 
     living_to_backyard = multiworld.get_entrance("Living Room - Backyard", player)
 
+    reach_second_floor = Has("Foyer Ladder") | Has("Foyer Teleport")
+
     foyer_to_jenny = multiworld.get_entrance("Foyer - Jenny's Room", player)
 
     foyer_to_bedroom = multiworld.get_entrance("Foyer - Bedroom", player)
@@ -51,9 +53,9 @@ def set_rules(self) -> None:
 
     self.set_rule(living_to_backyard, can_enter_backyard)
 
-    self.set_rule(foyer_to_jenny, can_enter_backyard)
+    self.set_rule(foyer_to_jenny, reach_second_floor)
 
-    self.set_rule(foyer_to_bedroom, can_enter_backyard)
+    self.set_rule(foyer_to_bedroom, reach_second_floor)
 
     has_all_items = HasAll(tooth_brush, blaster, charge_chip, squirter, mug, "Alien Ear Chip", "Giga-Battery", "Wedding Band", "Chibi-Radar Chibi-Gear")
 
@@ -79,37 +81,36 @@ def set_location_rules(self) -> None:
 
     # reach_jenny_battery = Has(red_shoe) & Has (blaster)
 
+    has_living_room_bridge_ladder = HasAll("Living Room Bridge", "Living Room Ladder")
+
     # Living Room
     living_room_frog_ring_window = multiworld.get_location("Living Room - Frog Ring (Behind Window)", player)
     self.set_rule(living_room_frog_ring_window, has_blaster)
 
-    # living_room_frog_ring_window = multiworld.get_location("Living Room - Plant Shelf Happy Block (Upper)", player)
-    # self.set_rule(living_room_frog_ring_window, has_blaster)
-    #
-    # living_room_frog_ring_window = multiworld.get_location("Living Room - Happy Block above Fireplace", player)
-    # self.set_rule(living_room_frog_ring_window, has_blaster)
-    #
-    # living_room_frog_ring_window = multiworld.get_location("Living Room - Happy Block above Chibi House", player)
-    # self.set_rule(living_room_frog_ring_window, has_blaster)
+    living_room_frog_ring_corkboard = multiworld.get_location("Living Room - Frog Ring (Corkboard)", player)
+    self.set_rule(living_room_frog_ring_corkboard, has_living_room_bridge_ladder)
 
     # Kitchen
-    kitchen_table_happy_block = multiworld.get_location("Kitchen - Spoon Location", player)
-    self.set_rule(kitchen_table_happy_block, has_blaster)
 
-    # kitchen_table_happy_block = multiworld.get_location("Kitchen - Table Happy Block", player)
-    # self.set_rule(kitchen_table_happy_block, has_blaster)
+    has_kitchen_ladder = Has("Kitchen Ladder")
 
-    # kitchen_cabinet_happy_block = multiworld.get_location("Kitchen - Cabinet Happy Block", player)
-    # self.set_rule(kitchen_cabinet_happy_block, has_blaster)
+    kitchen_spoon = multiworld.get_location("Kitchen - Spoon Location", player)
+    self.set_rule(kitchen_spoon, has_kitchen_ladder)
+
+    kitchen_cookie_crumb_by_spoon = multiworld.get_location("Kitchen - Cookie Crumbs by Spoon", player)
+    self.set_rule(kitchen_cookie_crumb_by_spoon, has_kitchen_ladder)
+
+    kitchen_cookie_box_a_by_spoon = multiworld.get_location("Kitchen - Cookie Box by Spoon A", player)
+    self.set_rule(kitchen_cookie_box_a_by_spoon, has_kitchen_ladder)
+
+    kitchen_cookie_box_b_by_spoon = multiworld.get_location("Kitchen - Cookie Box by Spoon B", player)
+    self.set_rule(kitchen_cookie_box_b_by_spoon, has_kitchen_ladder)
 
     kitchen_bandage = multiworld.get_location("Kitchen - Bandage Location", player)
-    self.set_rule(kitchen_bandage, has_blaster)
+    self.set_rule(kitchen_bandage, has_kitchen_ladder)
 
     kitchen_frog_ring_table = multiworld.get_location("Kitchen - Frog Ring (Table)", player)
     self.set_rule(kitchen_frog_ring_table, has_blaster)
-
-    # kitchen_high_cupboard_10_coin = multiworld.get_location("Kitchen - High Cupboard 10M Coin", player)
-    # self.set_rule(kitchen_high_cupboard_10_coin, has_blaster)
 
     #  Drain
     sink_frog_ring = multiworld.get_location("Sink Drain - Frog Ring", player)
@@ -118,9 +119,6 @@ def set_location_rules(self) -> None:
     #  Foyer
     foyer_frog_ring = multiworld.get_location("Foyer - Waterfall Frog Ring", player)
     self.set_rule(foyer_frog_ring, has_blaster)
-
-    # foyer_frog_ring = multiworld.get_location("Foyer - Top of Stairs 10M Coin", player)
-    # self.set_rule(foyer_frog_ring, has_blaster)
 
     #  Basement
     basement_giga_charger = multiworld.get_location("Basement - Giga Charger", player)
@@ -144,39 +142,9 @@ def set_location_rules(self) -> None:
     basement_cabinet_trash_b = multiworld.get_location("Basement - Cabinet Trash B", player)
     self.set_rule(basement_cabinet_trash_b, reach_charger)
 
-    # basement_shelf_happy_block_b= multiworld.get_location("Basement - Shelf Happy Block B", player)
-    # self.set_rule(basement_shelf_happy_block_b, reach_charger)
-    #
-    # basement_shelf_happy_block_a = multiworld.get_location("Basement - Shelf Happy Block A", player)
-    # self.set_rule(basement_shelf_happy_block_a, reach_charger)
-    #
-    # basement_rafters_happy_block_b = multiworld.get_location("Basement - Rafters Happy Block B", player)
-    # self.set_rule(basement_rafters_happy_block_b, reach_charger)
-    #
-    # basement_rafters_happy_block_a = multiworld.get_location("Basement - Rafters Happy Block A", player)
-    # self.set_rule(basement_rafters_happy_block_a, reach_charger)
-    #
-    # basement_swing_10m_coin = multiworld.get_location("Basement - Swing 10M Coin", player)
-    # self.set_rule(basement_swing_10m_coin, reach_charger)
-
     # Backyard
     # backyard_ship = multiworld.get_location("Backyard - Scurvy Splinter", player)
     # self.set_rule(backyard_ship, reach_ship)
-
-    # backyard_right_awning_happy_block_c = multiworld.get_location("Backyard - Right Awning Happy Block C", player)
-    # self.set_rule(backyard_right_awning_happy_block_c, reach_backyard_awning)
-    #
-    # backyard_right_awning_happy_block_b = multiworld.get_location("Backyard - Right Awning Happy Block B", player)
-    # self.set_rule(backyard_right_awning_happy_block_b, reach_backyard_awning)
-    #
-    # backyard_left_awning_happy_block = multiworld.get_location("Backyard - Left Awning Happy Block", player)
-    # self.set_rule(backyard_left_awning_happy_block, reach_backyard_awning)
-    #
-    # backyard_tree_happy_block = multiworld.get_location("Backyard - Tree Happy Block", player)
-    # self.set_rule(backyard_tree_happy_block, reach_backyard_tree_happy_block)
-    #
-    # backyard_right_awning_happy_block_a = multiworld.get_location("Backyard - Right Awning Happy Block A", player)
-    # self.set_rule(backyard_right_awning_happy_block_a, reach_backyard_awning)
 
     backyard_white_block= multiworld.get_location("Backyard - White Block", player)
     self.set_rule(backyard_white_block, reach_backyard_awning)